Packaging and CI (#2)
Create YAML for a basic workflow that tries to install the package and runs integration test.

Use `pyproject.toml` instead of `setup.py` for package metadata. 
This is the modern method for specifying package metadata; see
<https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html>
for further information.

Test plan: Installing to a fresh venv works as expected, and the version
number is populated correctly. Workflow runs as expected

Co-authored-by: YijieYin <yy432@cam.ac.uk>